requirements.txt
asgiref==3.6.0
black==22.12.0
bleach==5.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Django==3.2.16
django-appconf==1.0.5
django-compressor==4.3.1
django-nested-admin==4.0.2
django-simple-metatags==2.0.3
django-summernote==0.8.20.0
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.3
Pillow==9.3.0
platformdirs==2.6.2
python-monkey-business==1.0.0
pytz==2022.7
rcssmin==1.1.1
requests==2.28.1
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.3
tmdbv3api==1.7.7
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
webencodings==0.5.1
whitenoise==6.2.0
